kdbus.h: use proper kernel datatypes for variables crossing the user/kernel boundry
In other words, no uint64_t craziness, the kernel wants to use __u64
instead.

I left the "name" variables alone, as "char" should be safe for them.

+++ test/test-kdbus.c
@@ -30,7 +30,7 @@ static int connect_to_bus(const char *path)
 		fprintf(stderr, "--- error when saying hello: %d (\"%s\")\n", err, strerror(errno));
 		return EXIT_FAILURE;
 	}
-	printf("-- Our peer ID for %s: %lu\n", path, hello.id);
+	printf("-- Our peer ID for %s: %lu\n", path, (uint64_t)hello.id);
 
 	return fd;
 }
